V2.2.0. * Add PHPStan for static code analize. * Refactor backend code base to meet PHPStan static code analyze requirements and fix all issues that PHPStan found * Fix some security issues * Complete refactor of JS assets to resolve all leftover corner-cases.
