+++ classes/Infrastructure/Wordpress/Hooks/Actions/AddExtraFeesAction.php
@@ -5,13 +5,27 @@ declare(strict_types=1);
 namespace SamedayCourier\Shipping\Infrastructure\Wordpress\Hooks\Actions;
 
 use SamedayCourier\Shipping\Domain\CarrierConstants;
-use SamedayCourier\Shipping\Infrastructure\Woo\Services\WooSessionHandler;
+use SamedayCourier\Shipping\Domain\Ports\ChosenPaymentMethodReaderInterface;
+use SamedayCourier\Shipping\Infrastructure\Woo\Services\WooChosenPaymentMethodReader;
 use SamedayCourier\Shipping\Infrastructure\Wordpress\Handlers\TranslatorHandler;
 use SamedayCourier\Shipping\Infrastructure\Wordpress\Services\CarrierSettingsServiceProvider;
 
 final class AddExtraFeesAction extends AbstractAction
 {
     private const ACTION = 'woocommerce_cart_calculate_fees';
+
+    /**
+     * @var ChosenPaymentMethodReaderInterface
+     */
+    private ChosenPaymentMethodReaderInterface $chosenPaymentMethodReader;
+
+    /**
+     * @param ChosenPaymentMethodReaderInterface|null $chosenPaymentMethodReader
+     */
+    public function __construct(?ChosenPaymentMethodReaderInterface $chosenPaymentMethodReader = null)
+    {
+        $this->chosenPaymentMethodReader = $chosenPaymentMethodReader ?? new WooChosenPaymentMethodReader();
+    }
 
     /**
      * @return string
@@ -62,7 +76,7 @@ final class AddExtraFeesAction extends AbstractAction
             return false;
         }
 
-        $chosenPaymentMethod = (new WooSessionHandler())->getChosenPaymentMethod();
+        $chosenPaymentMethod = $this->chosenPaymentMethodReader->getChosenPaymentMethod();
 
         return CarrierConstants::CASH_ON_DELIVERY === $chosenPaymentMethod;
     }
